+++ src/main/java/br/com/extra/api/pojo/v2/orders/Freight.java
@@ -1,6 +1,7 @@
 package br.com.extra.api.pojo.v2.orders;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -11,25 +12,25 @@ import java.util.Date;
 public class Freight implements Serializable {
 
     private static final long serialVersionUID = -178398835901779098L;
-    private Double actualAmount;
-    private Double chargedAmount;
+    private BigDecimal actualAmount;
+    private BigDecimal chargedAmount;
     private String type;
     private Date scheduledAt;
     private String scheduledPeriod;
 
-    public Double getActualAmount() {
+    public BigDecimal getActualAmount() {
         return actualAmount;
     }
 
-    public void setActualAmount(Double actualAmount) {
+    public void setActualAmount(BigDecimal actualAmount) {
         this.actualAmount = actualAmount;
     }
 
-    public Double getChargedAmount() {
+    public BigDecimal getChargedAmount() {
         return chargedAmount;
     }
 
-    public void setChargedAmount(Double chargedAmount) {
+    public void setChargedAmount(BigDecimal chargedAmount) {
         this.chargedAmount = chargedAmount;
     }
